python/soupbot/requirements.txt
discord.py==2.0.1 